dom/tests/mochitest/bugs/test_bug504862.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=504862
-->
<head>
  <title>Test for Bug 504862</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runTest()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=504862">Mozilla Bug 504862</a>
<script class="testbody" type="text/javascript">

if (!navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(4);
}

/** Test for Bug 504862 **/
SimpleTest.waitForExplicitFinish();
function onMsgRcv(event)
{
  is(event.data, "args: undefined", "Unexpected cross origin dialog arguments.");
}

function runTest() {
  window.addEventListener("message", onMsgRcv, false);

  var result = window.showModalDialog("file_bug504862.html", "my args");
  is(result, null, "window sees previous dialog documents return value.");

  result = window.showModalDialog("http://test1.example.com/tests/dom/tests/mochitest/bugs/file_bug504862.html", "my args");

  is(result, null, "Able to see return value from cross origin dialog.");

  // Garbage collecting the windows created in this test can cause
  // assertions, so GC now to blame those assertions to this test.
  // ("mArguments wasn't cleaned up properly!" in ~nsGlobalWindow,
  // bug 600703)
  SpecialPowers.gc();

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
